+++ hexital/__init__.py
@@ -1,3 +1,3 @@
 from .analysis import movement, patterns
-from .core import OHLCV, Hexital
+from .core import Candle, Hexital
 from .indicators import *