+++ hexital/core/__init__.py
@@ -1,3 +1,3 @@
+from .candle import Candle
 from .hexital import Hexital
 from .indicator import Indicator
-from .ohlcv import OHLCV