+++ tests/utilities/test_utils.py
@@ -1,16 +1,22 @@
 import pytest
-from hexital.core.ohlcv import OHLCV
+from hexital.core.candle import Candle
 from hexital.lib.utils import candles_sum
 
 
 @pytest.fixture(name="rising_candles")
 def fixture_rising_candles():
     return [
-        OHLCV(open=90, high=110, low=80, close=80, volume=10, indicators={"ATR": 600}),
-        OHLCV(open=100, high=120, low=90, close=90, volume=10, indicators={"ATR": 700}),
-        OHLCV(open=110, high=130, low=100, close=100, volume=10, indicators={"ATR": 800}),
-        OHLCV(open=120, high=140, low=110, close=110, volume=10, indicators={"ATR": 900}),
-        OHLCV(open=130, high=150, low=120, close=120, volume=10, indicators={"ATR": 990}),
+        Candle(open=90, high=110, low=80, close=80, volume=10, indicators={"ATR": 600}),
+        Candle(open=100, high=120, low=90, close=90, volume=10, indicators={"ATR": 700}),
+        Candle(
+            open=110, high=130, low=100, close=100, volume=10, indicators={"ATR": 800}
+        ),
+        Candle(
+            open=120, high=140, low=110, close=110, volume=10, indicators={"ATR": 900}
+        ),
+        Candle(
+            open=130, high=150, low=120, close=120, volume=10, indicators={"ATR": 990}
+        ),
     ]